Fix how we deal with legacy services.
